chore: move UI components
