+++ src/features/products/Products.tsx
@@ -2,7 +2,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { getProducts } from '@/store';
 import type { Product } from '@/types';
-import ProductsTable from './ProductsTable';
+import ProductsTable from './ui/ProductsTable';
 
 export default function Products() {
   const { isLoading, isError, data, error } = useQuery<Product[], Error>({
